refactor: extract shared config types (TLS, Retry, Batch, Compression)
Add shared.rs with TlsClientConfig, TlsServerConfig, RetryConfig,
BatchConfig, and Compression enum. These provide a uniform config
surface for all inputs and outputs.

Rename TlsInputConfig -> TlsServerConfig (with backward-compat alias).
Export TlsClientConfig from the crate.

This is the foundation for #2194-#2200 (per-component config options).

Closes #2201
