+++ crates/logfwd-config/src/lib.rs
@@ -11,11 +11,15 @@ mod compat;
 mod env;
 mod load;
 mod serde_helpers;
+pub mod shared;
 mod types;
 mod validate;
 
 #[cfg(test)]
 pub(crate) use env::expand_env_vars;
+pub use shared::{
+    BatchConfig, Compression, RetryConfig, TlsClientConfig, TlsInputConfig, TlsServerConfig,
+};
 pub use types::{
     ArrowIpcTypeConfig, AuthConfig, Config, ConfigError, CsvEnrichmentConfig, EnrichmentConfig,
     FileTypeConfig, Format, GeneratorAttributeValueConfig, GeneratorComplexityConfig,
@@ -25,7 +29,7 @@ pub use types::{
     JournaldBackendConfig, JournaldInputConfig, JournaldTypeConfig, JsonlEnrichmentConfig,
     K8sPathConfig, OtlpProtobufDecodeModeConfig, OtlpTypeConfig, OutputConfig, OutputType,
     PipelineConfig, SensorTypeConfig, ServerConfig, StaticEnrichmentConfig, StorageConfig,
-    TcpTypeConfig, TlsInputConfig, UdpTypeConfig,
+    TcpTypeConfig, UdpTypeConfig,
 };
 pub use validate::validate_host_port;
